feat(springboot/java upgrade): renamed date service package (previous called locationservice) Refs #94
