+++ date-service/pom.xml
@@ -41,8 +41,8 @@
 
         <!-- tag::kafka[] -->
         <dependency>
-            <groupId>org.springframework.kafka</groupId>
-            <artifactId>spring-kafka</artifactId>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-kafka</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.avro</groupId>
@@ -197,12 +197,11 @@
                             <inputSpec>${project.basedir}/../specs/profiles.yaml</inputSpec>
                             <generatorName>java</generatorName>
                             <library>resttemplate</library>
+                            <generateModelTests>false</generateModelTests>
+                            <generateApiTests>false</generateApiTests>
                             <configOptions>
                                 <useJakartaEe>true</useJakartaEe>
                                 <interfaceOnly>true</interfaceOnly>
-                                <generateSupportingFiles>false</generateSupportingFiles>
-                                <generateModelTests>false</generateModelTests>
-                                <generateApiTests>false</generateApiTests>
                             </configOptions>
                         </configuration>
                     </execution>
@@ -215,12 +214,11 @@
                             <inputSpec>${project.basedir}/../specs/location.yaml</inputSpec>
                             <generatorName>java</generatorName>
                             <library>resttemplate</library>
+                            <generateModelTests>false</generateModelTests>
+                            <generateApiTests>false</generateApiTests>
                             <configOptions>
                                 <useJakartaEe>true</useJakartaEe>
                                 <interfaceOnly>true</interfaceOnly>
-                                <generateSupportingFiles>false</generateSupportingFiles>
-                                <generateModelTests>false</generateModelTests>
-                                <generateApiTests>false</generateApiTests>
                             </configOptions>
                         </configuration>
                     </execution>
