+++ location-service/src/main/resources/application.yaml
@@ -3,7 +3,6 @@ spring:
     default-property-inclusion: non_null
     serialization:
       FAIL_ON_EMPTY_BEANS: false
-      WRITE_DATES_AS_TIMESTAMPS: false
   web:
     error:
       include-message: always
